+++ BinaryLogic/Circle.cs
@@ -11,29 +11,29 @@ namespace BinaryLogic
     public class Circle
     {
         public Point position;
-        public float radius;
+        public int radius;
 
-        public Circle(Point pos, float radius)
+        public Circle(Point pos, int radius)
         {
             position = pos;
             this.radius = radius;
         }
 
-        public void Move(Direction direction, float units = 1)
+        public void Move(Direction direction, uint units = 1)
         {
             switch (direction)
             {
                 case Direction.Up:
-                    position.Y -= units;
+                    position.Y -= (int)units;
                     break;
                 case Direction.Down:
-                    position.Y += units;
+                    position.Y += (int)units;
                     break;
                 case Direction.Right:
-                    position.X += units;
+                    position.X += (int)units;
                     break;
                 case Direction.Left:
-                    position.X -= units;
+                    position.X -= (int)units;
                     break;
             }
         }
